refactor 잘못된 코드 수정
