--- toyProject/src/main/java/com/mamully/toyProject/config/Webconfig.java
@@ -1,25 +1,0 @@
-package com.mamully.toyProject.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class Webconfig {
-    @Configuration
-    public class WebConfig {
-
-        @Bean
-        public WebMvcConfigurer corsConfigurer() {
-            return new WebMvcConfigurer() {
-                @Override
-                public void addCorsMappings(CorsRegistry registry) {
-                    registry.addMapping("/api/**")
-                            .allowedOrigins("http://localhost:8084")  // 허용할 도메인 설정
-                            .allowedMethods("GET", "POST", "PUT", "DELETE");
-                }
-            };
-        }
-    }
-}
